Update and move the low level stuff
